requirements.txt
Flask
dotenv
azure-servicebus

